% cleanup
git-svn-id: http://svn.code.sf.net/p/htmlunit/code/trunk/htmlunit@6392 5f5364db-9458-4db8-a492-e30667be6df6

+++ src/main/java/com/gargoylesoftware/htmlunit/javascript/background/JavaScriptJob.java
@@ -112,7 +112,7 @@ public abstract class JavaScriptJob implements Runnable, Comparable<JavaScriptJo
         final boolean xhr2 = other.executeAsap_;
 
         if (xhr1 && xhr2) {
-            return getId() - other.getId();
+            return getId().intValue() - other.getId().intValue();
         }
 
         if (xhr1) {
